# apps/swagger/src/spec/messages.spec.yaml
openapi: 3.0.0
info:
  title: Meta Cloud API
  version: 1.0.0
  description: API documentation
servers:
  - url: http://localhost:3000
components:
  schemas:
    MessagesResponse:
      type: object
      properties:
        messaging_product:
          type: string
          enum: [whatsapp]
        contacts:
          type: array
          items:
            type: object
            properties:
              input:
                type: string
              wa_id:
                type: string
        messages:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
...
paths:
  /messages/audio:
    post:
      summary: Send an audio message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    id:
                      type: string
                      description: Required when not using link. ID of media object
                    link:
                      type: string
                      description: Required when not using id. HTTP/HTTPS URL of media
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/contacts:
    post:
      summary: Send a contacts message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: array
                  required: true
                  maxItems: 1
                  minItems: 1
                  items:
                    type: object
                    required:
                      - name
                    properties:
                      addresses:
                        type: array
                        items:
                          type: object
                          properties:
                            street:
                              type: string
                            city:
                              type: string
                            state:
                              type: string
                            zip:
                              type: string
                            country:
                              type: string
                            country_code:
                              type: string
                            type:
                              type: string
                              enum: [HOME, WORK]
                      birthday:
                        type: string
                        pattern: '^\d{4}-\d{2}-\d{2}$'
                      emails:
                        type: array
                        items:
                          type: object
                          properties:
                            email:
                              type: string
                            type:
                              type: string
                              enum: [HOME, WORK]
                      name:
                        type: object
                        required:
                          - formatted_name
                        properties:
                          formatted_name:
                            type: string
                          first_name:
                            type: string
                          last_name:
                            type: string
                          middle_name:
                            type: string
                          suffix:
                            type: string
                          prefix:
                            type: string
                      org:
                        type: object
                        properties:
                          company:
                            type: string
                          department:
                            type: string
                          title:
                            type: string
                      phones:
                        type: array
                        items:
                          type: object
                          properties:
                            phone:
                              type: string
                            type:
                              type: string
                              enum: [CELL, MAIN, IPHONE, HOME, WORK]
                            wa_id:
                              type: string
                      urls:
                        type: array
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                            type:
                              type: string
                              enum: [HOME, WORK]
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/document:
    post:
      summary: Send a document message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    id:
                      type: string
                      description: Meta media ID for the document
                    link:
                      type: string
                      description: URL of the document
                    caption:
                      type: string
                      description: Optional caption for the document
                    filename:
                      type: string
                      description: Optional filename for the document
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              withId:
                value:
                  body:
                    id: "123456789"
                    caption: "Document with ID"
                    filename: "report.pdf"
                  recipient: 821089791169
              withLink:
                value:
                  body:
                    link: "https://example.com/document.pdf"
                    caption: "Document with URL"
                    filename: "document.pdf"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/image:
    post:
      summary: Send an image message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    id:
                      type: string
                      description: Meta media ID for the image
                    link:
                      type: string
                      description: URL of the image
                    caption:
                      type: string
                      description: Optional caption for the image
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              withId:
                value:
                  body:
                    id: "123456789"
                    caption: "Image with ID"
                  recipient: 821089791169
              withLink:
                value:
                  body:
                    link: "https://example.com/image.jpg"
                    caption: "Image with URL"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/interactive:
    post:
      summary: Send an interactive message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    type:
                      type: string
                      enum: [button, list, product, product_list]
                      description: Type of interactive message
                    body:
                      type: object
                      required: true
                      properties:
                        text:
                          type: string
                    footer:
                      type: object
                      properties:
                        text:
                          type: string
                    header:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [document, image, text, video]
                        document:
                          type: object
                          properties:
                            id:
                              type: string
                            link:
                              type: string
                            caption:
                              type: string
                            filename:
                              type: string
                        image:
                          type: object
                          properties:
                            id:
                              type: string
                            link:
                              type: string
                            caption:
                              type: string
                        text:
                          type: string
                        video:
                          type: object
                          properties:
                            id:
                              type: string
                            link:
                              type: string
                            caption:
                              type: string
                    action:
                      type: object
                      required: true
                      properties:
                        button:
                          type: string
                        buttons:
                          type: array
                          items:
                            type: object
                            properties:
                              type:
                                type: string
                                enum: [reply]
                              reply:
                                type: object
                                properties:
                                  title:
                                    type: string
                                  id:
                                    type: string
                        catalog_id:
                          type: string
                        product_retailer_id:
                          type: string
                        sections:
                          type: object
                          properties:
                            product_items:
                              type: array
                              items:
                                type: object
                                properties:
                                  product_retailer_id:
                                    type: string
                            rows:
                              type: array
                              items:
                                type: object
                                properties:
                                  id:
                                    type: string
                                  title:
                                    type: string
                                  description:
                                    type: string
                            title:
                              type: string
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              button:
                value:
                  body:
                    type: button
                    body:
                      text: "Click one of the buttons"
                    action:
                      buttons:
                        - type: reply
                          reply:
                            id: "button1"
                            title: "Button 1"
                        - type: reply
                          reply:
                            id: "button2" 
                            title: "Button 2"
                  recipient: 821089791169
              list:
                value:
                  body:
                    type: list
                    body:
                      text: "Please select from the list"
                    action:
                      button: "Show List"
                      sections:
                        - title: "Section 1"
                          rows:
                            - id: "1"
                              title: "Option 1"
                              description: "Description 1"
                            - id: "2"
                              title: "Option 2"
                              description: "Description 2"
                  recipient: 821089791169
              product:
                value:
                  body:
                    type: product
                    action:
                      catalog_id: "123456"
                      product_retailer_id: "SKU123"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/location:
    post:
      summary: Send a location message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    longitude:
                      type: number
                      description: Longitude of the location
                    latitude:
                      type: number
                      description: Latitude of the location
                    name:
                      type: string
                      description: Optional name of the location
                    address:
                      type: string
                      description: Optional address of the location
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              basic:
                value:
                  body:
                    longitude: -122.4194
                    latitude: 37.7749
                  recipient: 821089791169
              withDetails:
                value:
                  body:
                    longitude: -122.4194
                    latitude: 37.7749
                    name: "San Francisco"
                    address: "San Francisco, CA 94105"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/sticker:
    post:
      summary: Send a sticker message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    id:
                      type: string
                      description: Meta media ID for the sticker
                    link:
                      type: string
                      description: URL of the sticker
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              withId:
                value:
                  body:
                    id: "123456789"
                  recipient: 821089791169
              withLink:
                value:
                  body:
                    link: "https://example.com/sticker.webp"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/template:
    post:
      summary: Send a template message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    name:
                      type: string
                      description: Name of the template
                    language:
                      type: object
                      properties:
                        policy:
                          type: string
                          enum: [deterministic]
                        code:
                          $ref: '#/components/schemas/LanguagesEnum'  
                    components:
                      type: array
                      items:
                        oneOf:
                          - type: object
                            properties:
                              type:
                                type: string
                                enum: [header, body, footer]
                              parameters:
                                type: array
                                items:
                                  oneOf:
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [text]
                                        text:
                                          type: string
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [currency]
                                        currency:
                                          type: object
                                          properties:
                                            fallback_value:
                                              type: string
                                            code:
                                              $ref: '#/components/schemas/CurrencyCodesEnum'
                                            amount_1000:
                                              type: number
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [date_time]
                                        date_time:
                                          type: object
                                          properties:
                                            fallback_value:
                                              type: string
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [image, document, video]
                                        image:
                                          type: object
                                          properties:
                                            link:
                                              type: string
                          - type: object
                            properties:
                              type:
                                type: string
                                enum: [button]
                              sub_type:
                                type: string
                                enum: [quick_reply, url]
                              index:
                                type: string
                              parameters:
                                type: array
                                items:
                                  oneOf:
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [payload]
                                        payload:
                                          type: string
                                    - type: object
                                      properties:
                                        type:
                                          type: string
                                          enum: [text]
                                        text:
                                          type: string
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              textTemplate:
                value:
                  body:
                    name: sample_text_template
                    language:
                      policy: deterministic
                      code: en_US
                    components:
                      - type: body
                        parameters:
                          - type: text
                            text: Hello {{1}}
                  recipient: 821089791169
              mediaTemplate:
                value:
                  body:
                    name: sample_media_template
                    language:
                      policy: deterministic
                      code: en_US
                    components:
                      - type: header
                        parameters:
                          - type: image
                            image:
                              link: https://example.com/image.jpg
                      - type: body
                        parameters:
                          - type: text
                            text: Product details
                  recipient: 821089791169
              interactiveTemplate:
                value:
                  body:
                    name: sample_interactive_template
                    language:
                      policy: deterministic
                      code: en_US
                    components:
                      - type: body
                        parameters:
                          - type: text
                            text: Would you like to proceed?
                      - type: button
                        sub_type: quick_reply
                        index: "0"
                        parameters:
                          - type: payload
                            payload: "YES_PAYLOAD"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/text:
    post:
      summary: Send a text message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    body:
                      type: string
                    preview_url:
                      type: boolean
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input
  /messages/video:
    post:
      summary: Send a video message
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: object
                  required: true
                  properties:
                    id:
                      type: string
                      description: Required when not using link. ID of media object
                    link:
                      type: string
                      description: Required when not using id. HTTP/HTTPS URL of media
                    caption:
                      type: string
                      description: Optional caption for the video
                recipient:
                  type: number
                  example: 821089791169
                replyMessageId:
                  type: string
            examples:
              withId:
                value:
                  body:
                    id: "123456789"
                    caption: "Video with ID"
                  recipient: 821089791169
              withLink:
                value:
                  body:
                    link: "https://example.com/video.mp4"
                    caption: "Video with URL"
                  recipient: 821089791169
      responses:
        200:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        400:
          description: Invalid input